+++ src/core/hub/index.ts
@@ -1,12 +1,10 @@
 import { ok, safeTry } from "neverthrow";
-import * as v from "valibot";
 
-import type { Executor } from "../executor";
 import type { NewDispatch, NewEvent } from "../model";
 import type { Repository } from "../repository";
 import type { EventPayload } from "../type";
 import { type QueueMessage, enqueue } from "./queue";
-import { Config, findRoutes } from "./routing";
+import { type Config, findRoutes } from "./routing";
 
 const constVoid = (() => {})();
 
@@ -80,76 +78,6 @@ export class EventSink {
 
     if (result.isErr()) {
       return Promise.reject(result.error);
-    }
-  }
-}
-
-export const getQueue = (env: Record<string, unknown>) => {
-  const queue = env.EVENTHUB_QUEUE;
-  if (!queue) {
-    throw new Error("cf-eventhub: EVENTHUB_QUEUE not set");
-  }
-  if (typeof queue !== "object" || "send" in queue) {
-    throw new Error("cf-eventhub: value of EVENTHUB_QUEUE is not a Queue");
-  }
-  return queue as Queue;
-};
-
-export const getRouteConfig = (env: Record<string, unknown>) => {
-  const routing = env.EVENTHUB_ROUTING;
-  if (!routing) {
-    throw new Error("cf-eventhub: EVENTHUB_ROUTING not set");
-  }
-  if (typeof routing !== "string") {
-    throw new Error("cf-eventhub: value of EVENTHUB_Routing is not a string");
-  }
-
-  return v.parse(Config, JSON.parse(routing));
-};
-
-export const getExecutor = (env: Record<string, unknown>) => {
-  const executor = env.EVENTHUB_EXECUTOR;
-  if (!executor) {
-    throw new Error("cf-eventhub: EVENTHUB_EXECUTOR not set");
-  }
-  if (typeof executor !== "object" || "dispatch" in executor) {
-    throw new Error(
-      "cf-eventhub: value of EVENTHUB_EXECUTOR is not a Executor",
-    );
-  }
-  return executor as Executor;
-};
-
-export class EventConsumer {
-  constructor(private executor: Executor) {}
-
-  private async dispatch(msg: Message<QueueMessage>) {
-    await this.executor
-      .dispatch(msg.body)
-      .then((result) => {
-        switch (result) {
-          case "complete":
-          case "ignored":
-          case "misconfigured":
-          case "notfound":
-            msg.ack();
-            break;
-          case "failed":
-            msg.retry();
-            break;
-          default: {
-            const _: never = result;
-          }
-        }
-      })
-      .catch(() => {
-        msg.retry();
-      });
-  }
-
-  async queue(batch: MessageBatch<QueueMessage>) {
-    for (const msg of batch.messages) {
-      await this.dispatch(msg);
     }
   }
 }
